requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
langchain
langchain-core
langchain-google-genai
langgraph
pydantic
